requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.6.1

# Azure Services
azure-ai-documentintelligence==1.0.0
azure-identity==1.15.0
openai==1.14.0

# Data processing
pandas==2.2.0
numpy>=1.26.0
scipy>=1.11.4

# Vector search
faiss-cpu>=1.8.0

# PDF processing
PyMuPDF>=1.24.0

# Image processing
Pillow>=10.0.0

# Web framework
fastapi==0.109.0
uvicorn==0.27.0

# HTTP client
aiohttp==3.9.1
httpx==0.25.2

# Async support
asyncio>=3.4.3

# Logging and monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2
